tar: split standalone AppleDouble test
