+++ libarchive/test/CMakeLists.txt
@@ -201,6 +201,7 @@ IF(ENABLE_TEST)
     test_read_format_tar_filename.c
     test_read_format_tar_invalid_pax_size.c
     test_read_format_tar_mac_metadata.c
+    test_read_format_tar_mac_metadata_standalone.c
     test_read_format_tar_pax_g_large.c
     test_read_format_tar_pax_large_attr.c
     test_read_format_tar_pax_negative_time.c
